Bug#38530445: recent regression: crashing in               calculate_materialization_costs
GROUPING function is now allowed to be present in a query if it
is explicitly grouped. It's no more required that ROLLUP needs to be
present in the query. For a subquery which is explicitly grouped,
GROUP BY could be removed as a redundant clause provided some conditions
are met. For the failing case, these conditions were met and therefore
was removed. This lead to missed checks for the arguments of GROUPING
function to be present in GROUP BY clause leading to problems later.

We now check for the presence of GROUPING function before removing
GROUP BY clause.

Change-Id: Ie82b299c246c45265aef69f1835711b312e69a32