Bug#38481979: Additional diagnostics on server-upgrade if table-opening fails
On server upgrade, we look at all table definitions, and if they
contain SQL (as part of constraints, etc.), we'll attempt to open
the table. If this fails (possibly because the SQL is no longer
valid in the new server, or because the sql_mode has changed, etc.),
we will inform the user of the issue (added by patch for Bug#36890891).

The patch adds the following:

- the table definition now contains the storage engine

- any and all error messages thrown during the aforementioned
  check are now printed to the error log as warnings, rather than
  elided.

- table-definitions are now emitted as warnings rather than notes

Change-Id: Ic1e961b0e882faa3877337209cb7350ca584ef7d