Bug#38465178: MySQL server 9.4.0 crashes at Item_func::walk
When finding replacements to handle materialized items for a window
function, hidden flag was not copied from the expression that was
being substituted. This led to problems when refreshing the ref
item slice. The field that a aggregate reference needs to point
to was wrongly picked as itself since this is based on the hidden
fields count(compute_ria_idx). So the call to ::walk() on this
expression would loop forever leading to server exit. We correct
this now.

Change-Id: I617ce78849befe0c0d0d2abde65d15d65a253694