Bug#38550210: Assertion failed: other_type != enum_json_type::J_ERROR
The issue here is a repeated call to an argument function of
the BENCHMARK() function, which fails with an assert in the second
invocaton of the benchmarked function. The problem is that the
benchmarked function sets an error status, and this is trapped in
the second invocation.

This is actually a legacy issue with BENCHMARK(). The problem is
revealed because the code that generates a JSON object from a TIME
value has recently been changed so that it no longer checks
THD::is_error(). This call prevented us from seeing the problem.

The solution is to add error handling to the BENCHMARK() function,
so that it bails out whenever an error is discovered. This is a
reasonable change, since calling an arbitrary function with an
error status has undefined behavior. The obvious sideeffect is
however that the BENCHMARK function will fail to invoke the
benchmarked function the specified number of times, if an error
is reported.

Change-Id: If1197415a678030908ac58dd39429f1a5003da13