BUG#38553372: innodb.monitor_restart fails on PB2
Test innodb.monitor_restart was failing on PB2 after the push of
WL#8602: "Make GTID_MODE = ON default".
The test has a debug point that asserts that the table used on the
test is the one that goes through a code path.
When gtid_mode is enable there is a system thread,
Clone_persist_gtid, that periodically updates the table
"mysql.gtid_executed". These updates go through the same code path
and was making the debug point assert to fail.

To avoid interfere with the test debug point, this test now runs
with gtid_mode disable.

Change-Id: I4eb6852588843438fa32b4ff2cdd9102a068532f