Bug#38328780: mysqld fails to start with assertions
Background:
-----------
When server is started with --skip-grant-tables privileges are not
checked. Remote connections are not allowed. In Windows, server has to
be run with TCP/IP, --shared-memory, or --named-pipe should be enabled
to allow remote connections. Otherwise server is not started and exits
execution early with a message that it was started with wrong
parameters.

Problem:
--------
During early return server did not end the replication tracker thread
and compress GTID thread before it started to shut down the plugins as
it is done during normal server shutdown. As a result of that InnoDB
shutdown failed because it detected open transactions from those
threads.

Fix:
----
During early exit caused by missing TCP/IP configuration,
--shared-memory, or --named-pipe parameters when using
--skip-grant-tables server stops replication tracker worker thread and
compress GTID thread, before shutting down plugins.

Change-Id: Ice534a5726adbe8e09555312396f0a7a0a813017