Bug#38328780: mysqld fails to start with assertions
Problem:
--------
In addition, replication tracker worker thread logic did not anticipate
thread to be stopped even before it started execution. This lead to
sporadic failures during early shutdown of the server because
THD_EXIT_COND was called on a thread even when it was not waiting on
condition.

Fix:
----
Replication tracker worker thread can be stopped even before
it was started without errors about incorrect condition exit check.

Change-Id: Ice534a5726adbe8e09555312396f0a7a0a813017