Bug#38430539: In some cases MDLs are released before Auto_releaser has run
Problem: Compiling with EXTRA_DD_DEBUG revealed that MDLs are sometimes
released before DD objects have been released.

Solution:
Remove EXTRA_DD_DEBUG flag and run these asserts in all debug builds.
It turns out that the process of creating an MDL_key for the checking is
costly and disruptive (requires uncached read of DD objects).
To mitigate this we keep MDL keys for acquired objects stored inside
each Auto_releaser, and check them both when the object is acquired and
when it is released by the Auto_releaser destructor.

Fix places where enabling these asserts revealed that locks were released too early,
typically by introducing another scope for the Auto_releaser.

One of the places was in uprade code which is not currently covered by
mtr tests, so here we add an 8.4 version of mysql_inplace_upgrade.test.

Change-Id: Ia3784e5b854be0adacaedc54c8b63d0ffd09cf23