Bug#38556362: Assertion failed: is_nullable(), file item_geofunc.cc
The problem here is an assert failure in a function that evaluates
an item that is not nullable, but nevertheless gets a NULL value
from the evaluation. The item that is evaluated is an IN predicate
with a TIME value on the left-hand side and two constant values on
the right-hand side. The data types are such that the bisection
algorithm cannot be used, hence evaluation is carried out by
cmp_item_datetime::cmp() which calls get_time_internal().

get_time_internal() tries to convert a string constant into a time
value, however the conversion fails, which generally would cause
a NULL value to be returned. But an IN predicate processes NULL values
by ignoring the NULL status and rather return -1, which is not a valid
TIME value, which would cause a false result. Let's call this a
"suppressed" NULL value.

cmp_item_datetime::cmp() can handle both "real" and "suppressed" NULL
values. Real NULL values are indicated by having item->null_value = true,
while "suppressed" NULL values are supposed to only set the is_null
output argument. get_time_internal() sets both these two in case of
a "suppressed" NULL value, which is not the proper behavior here.

Problem is fixed by only returning is_null = true when encountering
a "suppressed" NULL value.

Change-Id: Id83a0d96c9db3f0d98ed3875db629a21ada9185d