WL#16937  InnoDB : Container Aware Startup option [1/2]
Description:
  This worklog introduces a boolean startup option container_aware. It
is a GLOBAL, READ_ONLY variable which defaults to OFF. When enabled, the
MySQL server will adhere to the resource limits set by the cgroup. When
disabled, the server will read the resource limits directly from the
host.

  A warning is emitted if this variable is disabled, but cgroup resource
limits are found. The server exits with an error message if this
variable is enabled and either the server is not running in a cgroup, or
the resource limits could not be discovered - due to bad permissions or
missing cgroup configuration files etc.

  The two functions `my_physical_memory` and `my_num_vcpus` are used by
the server to find the amount of memory and number of logical CPUs
available to the server. They are updated to read the cgroup limits only
if container_aware is enabled.

  This introduces a depenedency: these functions cannot be called until
the value of the sysvar is set based on the server configuration. Thus,
asserts are added to ensure that these functions are called only after
the module has been initialized.

Additional Notes:
  WL#16484 had explicitly set the default at the time of declaring the
sysvar. The sysvar was declared as a C++ static variable, which is
initialized even before the main function is called. In order to work
with the dependency of the startup variable, these defaults are set to
a placeholder value. Later on, during plugin initialization; the
defaults are recalculated and the variable value is updated only if it
is not explicitly configured by the user.

  Similarly, temptable_max_ram is updated during server startup.

  Initialization of the resource groups also depended on `my_num_vpucs`.
This is moved after the container_aware variable is initialized (ie,
after init_common_variables in mysqld_main).

Change-Id: I2de20caca6d2f843bc86ab88142a23c91e499be2