Bug#38557567: Assertion `IsEmpty(child.delayed_predicates)' failed|join_optimizer/cost_model.cc [2/2, fix]
When using the hypergraph optimizer, an assertion in
EstimateStreamCost() fails when a STREAM access path is added on top
of an index nested loop join in which the join predicate is subsumed
by the index lookup.

The failing assertion checks that there are no delayed predicates to
apply after the STREAM path. This assumption holds, but the check for
it is incorrect. It checks that the AccessPath::delayed_predicates
bitset is empty. However, the delayed_predicates bitset serves two
purposes and consists of two partitions. The first
num_where_predicates bits represent delayed predicates, and the bits
beyond num_where_predicates represent subsumed sargable join
predicates.

In the failing case, the assertion sees the subsumed sargable join
predicates in the access path and mistakes them for delayed
predicates, and fails.

The fix is to make the assertion check only the first partition of
delayed_predicates.

The assertion is also moved from EstimateStreamCost() to
CreateMaterializationOrStreamingPath(), since it has nothing to do
with cost estimation.

Change-Id: I91fd50fab87fe6fcfb2c8d367b1f53724b1f6573