Bug#38557567: Assertion `IsEmpty(child.delayed_predicates)' failed|join_optimizer/cost_model.cc [1/2, cleanup, noclose]
This patch performs some initial cleanup of the cost estimation code
for aggregation, streaming and materialization. It does not fix the
bug.

1) The cost estimation functions EstimateAggregateCost() and
EstimateStreamCost() initialize some access path properties that are
not directly related to the cost estimates. These properties are
ordering_state and safe_for_rowid. Move the initialization of these
properties to where the access paths are created, so that the cost
model code doesn't touch unrelated properties of the access paths.

2) Move NewStreamingAccessPath() out-of-line, since it's too big and
too infrequently called to warrant inlining.

3) Some functions in sql_executor.cc for estimating stream and
materialization cost used current_thd even though the THD was
available directly. Change them to use thd instead of current_thd.

Change-Id: I4e2a01f5a8d3f1fb4c8ab34799ad68420371e230