Bug#36527984: Wrong authentication error message after disabling mysql_native_password
When an unknown user account tries to login the server picks a plugin at
random to complete the authentication exchange with and only then deny
the login.

It's picking a plugin at random from the list of "built-in" plugins,
  i.e. the plugins that are compiled into mysqld.

But it was not checking if the plugin selected is also enabled.
And, for some built-in plugins one can still have them loaded, but
disabled.
This was causing a different error message (plugin not loaded) vs the
expected error message: login denied.

Fixed by filling up a list of enabled plugins at startup and then
having the random pick out of these instead of all of the built-in
plugins.

Test added.

Change-Id: I437db95b472822522102c2f403735594110c2261