Bug#38500023 SUMA sends BLOB data with NO-LOGGING set [9.4+ test]
While the feature of sliced binlog epochs does not apply to BLOBs,
this commit adds a 9.4+ test to assess that the feature filters out
data at the NDB node, and there should not exist more byte counting
than necessary for a NDB API node that subscribes to the table.

The motivation is that byte counting for Ndb_api counters occurs at
the receipt of the data. While it could be assessed that the binlog
injector only wrote the row slice, there could still exist data coming
into the NDB API but un-merged and eventually discarded (like
Bug#38500023 resolution showed).

Change-Id: I06ff910a7abf5f3dc2ff08726e057647f4044128