Bug#38556168: extractvalue: Assertion !null_value || is_nullable failed
The problem is a failed assert due to an item evaluation indicating
a NULL value from a non-nullable item. The item that was evaluated was
an Item_func_equal. Even though it is never NULL, processing requires
that it is temporarily set as NULL. However, in the error case, this
null value is never cleared, thus we end in the reported assert.

The fix is to clear the null value in case of an error, to ensure that
a non-nullable item never returns with a null value.

A problem with the MTR test main.cast was revealed: A test case
performed "set names latin1" but never restored the default.
This problem was fixed, along with some necessary adjustments for
subsequent test cases.

Change-Id: I29d1e120feef917cd14bd1fb43d1ee2b25b65a91