Bug#38465147: MySQL server 9.4.0 crashes at               Item_view_ref::check_column_privileges
When re-resolving a const expression in an order by clause
that is also merged derived expression, things go wrong
because the name of the merged derived table and the
derived table that this order by was merged into are the
same.
Problem is not seen with non-const expressions because
we do not re-resolve such expressions. We detect that
if an order by expression is already resolved, then it
should be from a merged derived table and therefore does
not need re-resolving.
We now check if const expressions are from a merged
derived table and follow the same rules that apply for
non-const expressions.

Change-Id: Iac79e282784665a947e547c4b70cfb55d5a96a59