Bug#38131344: mysqld crash : join_optimizer , subgraph_enumeration, {anonymous}::CostingReceiver::ProposeAccessPath
The hypergraph optimizer raised assert failures for some joins which
had sargable join predicates with very low selectivity.

The problem was some code in ApplyDelayedPredicatesAfterJoin() which
compensated for double-counted selectivity, but only if the
selectivity was greater than 1e-6. So for example for join predicates
on primary keys in tables with more than one million row, the
double-counting was not compensated for. This was presumably done in
order to prevent division by zero.

Fixed by enabling the compensating code for all predicates that have
selectivity different from zero.

Change-Id: Iaf87aacc8736525fdb7db40d3c4dc02b69e76b69