Bug#38608041: Hypergraph creates temporary AccessPathSet objects unnecessarily
CostingReceiver::ProposeAccessPathWithOrderings() retrieves the access
path set for the current set of tables from an unordered map, or
inserts an empty set if there is no existing set for those tables. It
uses std::unordered_map::emplace() for that, and it always creates an
AccessPathSet object, regardless of whether one should be inserted or
not.

Creating and destroying these objects takes time, and doing it once
per subplan considered by the hypergraph optimizer could add up and
become significant during planning of complicated queries.

The solution is to use std::unordered_map::try_emplace() instead. That
function constructs an AccessPathSet only if there is no existing set
in the map, so we avoid the construction and destruction cost when the
map already has a set for the tables of the subplan.

This shaves off roughly 8% of the planning time for TPC-DS Q64.

Change-Id: I4c29b9727852901babf1224b51fc8d739640d5e3