Bug#38612660: Hypergraph: Cache reachable nodes in AccessPathSet
Profiling shows that FindReachableTablesFrom() takes about 8% of the
planning time for a complex query such as TPC-DS Q64. Caching the
calculated value in AccessPathSet reduces the time spent so much the
function doesn't show up in the profile any more, and the planning
time is reduced by about the same amount.

Change-Id: I571bcb7a92d9db0e201fc2cd17ef43e3e35365c6