Bug#38618473: Improve diagnostics messages for temporal value overflow
Warnings and errors issued when the value calculated in a temporal
function overflows could be more helpful.

This is an offending statement and its corresponding warning:

  SELECT CAST('00:00:00' AS TIME) + INTERVAL 2000 HOUR:

  Warning | 1441 | Datetime function: datetime field overflow

We are adding a new diagnostics code ER_TEMPORAL_FUNCTION_OVERFLOW that
can be used to provide better warnings and errors for such cases, thus
this is the new warning:

  Warning | 6527 | TIME value is out of range in
                   '(cast('00:00:00' as time) + interval 2000 hour)'

The format of the message is modelled after the corresponding message
for numerical overflow ER_DATA_OUT_OF_RANGE. Note however that when
ER_DATA_OUT_OF_RANGE is used, an unconditional error is issued, whereas
in the case of a temporal overflow, a warning is issued in non-strict
mode and an error is issued in strict mode. This is done in order to
preserve current semantics as much as possible.

Change-Id: Ib5c2c4ce7279c4d9623780d17d21945ab6307511