Bug#38625494: Incorrect error handling in open_table_uncached()
The code in open_table_uncached() which cleans up after errors raised
by open_table_def(), attempts to destruct a TABLE object that has not
yet been constructed, which yields undefined behaviour.

The TABLE object in question is allocated, and needs to be freed, but
it is not constructed until open_table_from_share() is called further
down in open_table_uncached().

The fix is to remove the call to the destructor.

The patch also adds code to fill the allocated memory with garbage and
inject an error, so that the error handling can be tested. It is hard
to test without error injection currently, because open_table_def()
usually doesn't raise an error unless it is running out of memory or
the database is corrupted.

Change-Id: I466dfdda3fd659e3944e9eb1411d6abd711d9a42