WL#16076 step 5: meta library: make Is_specialization a concept
WL#16076: New GTID set data structure

Change `Is_specialization` to be a concept, rather than a type. This
simplifies the syntax for using the tool from `Is_specialization<A,
B>::value` to `Is_specialization<A, B>`. It also enables use of
`Is_specialization` where a concept is required, such as function
prototypes.

Also, introduce a `Is_nontype_specialization`, which works for
templates with non-type arguments.

Change-Id: I81d3d4231ff986f51a7935794acbd77e2dcb1f32