WL#16441 Bulk load into non-empty table
- This WL lifts the restriction of bulk load only being available on empty tables.
To do this, we implement a table reader object that reads the records from the
existing table for all indexes. We instantiate parallel readers based on key
ranges that we get either by reading the first record in each CSV file during the
non-sorting load, or by getting the partitioning keys during the sorting load. We
create a duplicate table with a temporary name which we use as the bulk load
target. The sub-tree loaders will use their respective table readers to perform
something like a merge sort with the existing table. We compare the records coming
in from the CSV input with the records coming from the existing table to insert
the existing records in the appropriate place in the input stream. When bulk load
successfully finishes, the existing table is dropped and the duplicate table is
renamed to match the original table name. In the particular case of ROW_ID based
clustered indexes we skip over the ROW_ID range contained in the original table
and use new ROW_IDs for the CSV input, while giving each worker a part of the
original ROW_ID range to load as a separate sub-tree.
- Changes to community dirs.

Change-Id: I57837c065540793d5cb34219861a0928b6efa039