Bug#38466946 - mysqld crash - Segmentation fault with stored function
Issue: When looking at a having clause, parser creates items of type
Item_ref for all columns. These Item_refs are ready only when they are
resolved which happens when fixing the HAVING caluse. Since we have a
stored function in the having clause, we are trying to inline it before
the argument (Item_ref in this case) is fixed. Hence we encounter a
nullptr.

Solution: In the given case we cannot use const_item() function.
We need to use basic_const_item() instead.

Change-Id: I0b59d955c7f2bc97d5a61a9fd5562c0ee0d0e9fb