Bug#37922543 Please add source node ID to log message
This bug report asks for more information in the message
"Transporter # to node # disconnected in state: #"

The patch adds the reporting node id to the beginning of the
message ("At node #, transporter ..."). Also, rather than reporting
a numeric state code (0 for connected / 1 for connecting), the new
message includes text "(was connecting)" only in the less common
case of disconnection during the CONNECTING state.

Change-Id: I1bb2d3a51b6b9a023bcdabd5d2b11ba2b102af83