Bug#38077617: MySQL 8.4 always uses caching_sha2_password in the initial connection handshake
In MySQL 8.4, the default authentication plugin is configured through
the "authentication_policy" parameter.

Even when the policy is configured as "*:mysql_native_password", the
authentication method in the initial handshake packet sent by the
server will still use "caching_sha2_password" plugin, this can create
issues for old clients that do not support caching_sha2_password
nor Protocol::Auth.

The fix makes sure that the default authentication plugin in the
initial handshake is the one taken from the policy itself, if candidate
exists (instead of always using the hardcoded default).

Change-Id: I13a08f88c37a9324208406e28e25f13bba403c2c