Bug#38701618: Some upstream tests fail against amd64v3 architecture variant
The main.subquery_sj_mat_bka_nobnl test failed in mysql-8.0 and
mysql-8.4 when compiled with -march=x86-64-v3 because recorded cost
estimates were slightly different. This was seen because compiling for
that architecture enables the FMA instruction, which could affect the
result of floating-point calculations.

Fixed by filtering out cost estimates.

This patch is for mysql-trunk, which already filtered out the cost
estimates from the statement that got different estimates depending on
compilation options. The patch filters out the remaining cost
estimates in the test for good measure.

Change-Id: Ia638bf6423eeeb238076d87176d9c52a015e7f4d