Bug#38208188: Bulk inserts into temporary tables with GIS indexes will inevitably cause crashes
Problem:
Bulk inserts into temporary tables with GIS indexes will inevitably
cause crashes.

Cause:
When an mtr starts, the default log_mode is MTR_LOG_ALL.

For temporary tables, the mtr->log_mode is set to MTR_LOG_NO_REDO,
because if redo log recovery logic sees a redo log record for a given
space_id, it will try to open the space file and apply it, which is
impossible for temporary tables.

While inserting an entry in secondary index, we first try with
optimistic descent and if it returns with DB_FAIL we opt for
pessimistic descent.

While optimistic tree descent, we pick BTR_MODIFY_LEAF as the
descent-mode. However if there is a need for enlarging the mbr we must
set the descent-mode to BTR_MODIFY_TREE. So the ongoing mtr is
commited, and a new mtr is started, but this mtr's log_mode is the
default MTR_LOG_ALL, causing an assertion.

Fix:
After the new mtr is started, if the table is temporary, re-set the
`log_mode` to `MTR_LOG_NO_REDO`.

Change-Id: I6b0a83ff3334411b7881c3b092bbde1c022d4bc0