Bug#38330571 cmake/abi_check.cmake stalls for a minute on Windows 11 24h2
Problem:

The cmake/abi_check.cmake aims to check if we are on Windows, with WSL,
with gcc, which can be used for ABI checks.
To check for gcc, it tries to execute `wsl which gcc` command, but
first makes sure that the `wsl` command itself is available.
Turns out, that at least since Windows 11 24h2, even if the Windows
Subsystem for Linux is not installed, there's still a stub wsl.exe
program, which when run with `wsl which gcc` prints:
"""
The Windows Subsystem for Linux is not installed.
You can install by running 'wsl.exe --install'.
For more information please visit https://aka.ms/wslinstall

Press any key to install Windows Subsystem for Linux.
Press CTRL-C or close this window to cancel.
This prompt will time out in 60 seconds.
"""
which causes cmake to stall for 60 seconds.

Solution:

Before attempting `wsl which gcc` that can block, first try to
execute `wsl --status` which returns immediately, with the return code
indicating if WSL is installed or not.

Change-Id: Ib5978746541cd74cd722417d9e12bb6d3fb95a7c