Bug#38624183 ndb_restore should optionally hint transaction coordinator
ndb_restore is modified to take a new optional parameter
  --hint-node=0|1 (Default 0)

When set to 0, there is no effect, and ndb_restore will continue
to use Round-Robin via optimized_node_selection et al to choose
a transaction coordinator.

When set to 1, ndb_restore will hint all transactions used for
restore of DATA and LOG file content towards the data node
whose backup filesets it is restoring.

Setting hint-node to 1 is expected to be performance neutral
or beneficial in all cases except when restoring a backup to
a cluster of a different topology, in which case it may reduce
restore performance.

MTR test ndb_restore_debug is modified to show the effect
of using hint-node.

Change-Id: I8e8da4dc688d22da1f7b0df83d561cc45afb4a3b