Bug#38456164: On MDS, backtraces from the traditional error are not always available in the other (e.g. JSON) logs
Due to an incompatability between the standard version of
a function and our portable version thereof, backtraces
in error log files were sometimes not processed as expected.
This led to those backtraces not appearing in the table,
performance_schema.error_log, nor in error log files other
than the traditional one (e.g. the information would not
be included in the JSON log in such cases).

Change-Id: I1f8828f93008d30b649adaa3f589c6077c743726